+++ Core/Src/fram.c
@@ -175,8 +175,8 @@ void framWriteEnable ( void )
  *  NOTE    : argument len must not be greater than 256
  *
  */
-void framReadMemory ( unsigned short addr, unsigned char* rdBufP,
-                      unsigned short len )
+void framReadMemory ( uint32_t addr, unsigned char* rdBufP,
+                      uint32_t len )
 {
     unsigned char txbuf [ 3 ] ;
 
@@ -210,8 +210,8 @@ void framReadMemory ( unsigned short addr, unsigned char* rdBufP,
  *  NOTE    : argument len must not be greater than 256
  *
  */
-void framWriteMemory ( unsigned short addr, const unsigned char* const wrBufP,
-                       unsigned short len )
+void framWriteMemory ( uint32_t addr, const unsigned char* const wrBufP,
+                       uint32_t len )
 {
     unsigned char txbuf [ 3 ] ;
     unsigned char sr ;
@@ -253,7 +253,7 @@ uint8_t framTest ( void )
     unsigned char txbuf [ TLEN ] ;
     unsigned char rxbuf [ TLEN ] ;
 
-    int i, pass ;
+    uint8_t i, pass ;
 
     /* initialize txbuf to incrementing pattern */
     for ( i = 0; i < TLEN; i++ )
